I refactored several of the methods in WidgetPackageUtils in the Parser sub-project into a new WidgetFileUtils class in the main project. The reason for this is that these methods are not used internally in the parser at all, and are utility methods used by the Wookie server for uploading, moving and deleting widget files.
Moving these methods also removes two dependencies from the parser, which is good as it should be as portable as we can make it. 

I also updated the file upload to use the latest methods in the FileUpload package as the one we were using had been deprecated; I've tested this out and it looks like it works OK, but please do try it out and see if there are any gotchas. Finally I added more tests for the remaining WidgetPackageUtils methods.

git-svn-id: http://svn.apache.org/repos/asf/incubator/wookie/trunk@935245 13f79535-47bb-0310-9956-ffa450edef68